refactor: rename modules to noun-oriented names
I've realized that it's perhaps not best to going against the grain
too much with respect to naming modules.

Previously modules were verb-oriented, which is fairly likely to strike
many developers as unfamilar enough to incite some level of
irritability

What's more is that, even in my experience, I find that naming modules according to
modules is more appropriate for more contexts than verb-oriented naming.

The very instance that provided fertile grounds for this renaming is

commit 14007fe

I realized when making this commit that my mind would feel more
comfortable with calling the FontChanger module, as opposed to
ChangeFont module