fix: remove check for database in client-side code
I intially wrote this portion of code fully knowing that the
sqlite3.connect method already creates the database if it does not
already exist.

However, I am one that favors transparency for the sake of clarity, even
if that transparency comes at the cost of being less concise. In the
context of this particular situation, however, my preference
unfortunately backfired as it ironically made the logic intractable
to deal with.